# frontend/public/demo-api/openapi.json
{
  "components": {
    "schemas": {
...
      "CategorySummary": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "count"
        ],
        "title": "CategorySummary",
        "type": "object"
      },
...
      "CompanySummary": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "count",
          "category"
        ],
        "title": "CompanySummary",
        "type": "object"
      },
...
      "DaySummariesResponse": {
        "properties": {
          "month": {
            "title": "Month",
            "type": "string"
          },
          "summaries": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Summaries",
            "type": "object"
          }
        },
        "required": [
          "month",
          "summaries"
        ],
        "title": "DaySummariesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JournalDay": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "day_total": {
            "title": "Day Total",
            "type": "number"
          },
          "mtd_total": {
            "title": "Mtd Total",
            "type": "number"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "date",
          "day_total",
          "count",
          "mtd_total",
          "transactions"
        ],
        "title": "JournalDay",
        "type": "object"
      },
      "JournalResponse": {
        "properties": {
          "budget_ceiling": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Ceiling"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/JournalDay"
            },
            "title": "Days",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "month_total": {
            "title": "Month Total",
            "type": "number"
          },
          "transaction_count": {
            "title": "Transaction Count",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "days",
          "month_total",
          "transaction_count",
          "budget_ceiling"
        ],
        "title": "JournalResponse",
        "type": "object"
      },
...
      "MonthPaceInfo": {
        "description": "Mid-month pace context \u2014 non-null only when the requested month is the\ncurrent month (see the summary router). Day *position* for every fraction\nlookup is ``day_of_month / days_in_month``.\n\n``breakdown`` is the commitment-aware decomposition (L6): non-null only when\nthe upcoming-charge derivation succeeds AND penciled charges exist; it\nfails open to ``None`` (curve-only pace) exactly like ``pace`` itself.",
        "properties": {
          "breakdown": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonthPaceBreakdown"
              },
              {
                "type": "null"
              }
            ]
          },
          "day_of_month": {
            "title": "Day Of Month",
            "type": "integer"
          },
          "days_in_month": {
            "title": "Days In Month",
            "type": "integer"
          },
          "forecast_quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Quality"
          },
          "previous_to_date": {
            "title": "Previous To Date",
            "type": "number"
          },
          "projected_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Lower"
          },
          "projected_month_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Month Total"
          },
          "projected_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Upper"
          },
          "typical_to_date": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical To Date"
          }
        },
        "required": [
          "day_of_month",
          "days_in_month",
          "previous_to_date",
          "typical_to_date",
          "projected_month_total",
          "projected_lower",
          "projected_upper",
          "forecast_quality"
        ],
        "title": "MonthPaceInfo",
        "type": "object"
      },
...
      "MonthSummary": {
        "examples": [
          {
            "by_category": {
              "groceries": {
                "amount": 612.45,
                "count": 18
              },
              "restaurants": {
                "amount": 248.9,
                "count": 11
              }
            },
            "by_company": {
              "WHOLE FOODS": {
                "amount": 412.3,
                "category": "groceries",
                "count": 8
              }
            },
            "deposit_count": 2,
            "deposit_total": 5200.0,
            "deposits_by_company": {
              "EMPLOYER PAYROLL": {
                "amount": 5200.0,
                "count": 2
              }
            },
            "spending_count": 84,
            "top_categories": [
              {
                "amount": 612.45,
                "count": 18,
                "name": "groceries"
              },
              {
                "amount": 248.9,
                "count": 11,
                "name": "restaurants"
              }
            ],
            "total_spending": 3142.87,
            "year_month": "2026-04"
          }
        ],
        "properties": {
          "by_category": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "title": "By Category",
            "type": "object"
          },
          "by_company": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CompanySummary"
            },
            "title": "By Company",
            "type": "object"
          },
          "deposit_count": {
            "title": "Deposit Count",
            "type": "integer"
          },
          "deposit_total": {
            "title": "Deposit Total",
            "type": "number"
          },
          "deposits_by_company": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DepositSourceSummary"
            },
            "title": "Deposits By Company",
            "type": "object"
          },
          "spending_count": {
            "title": "Spending Count",
            "type": "integer"
          },
          "top_categories": {
            "items": {
              "$ref": "#/components/schemas/TopCategory"
            },
            "title": "Top Categories",
            "type": "array"
          },
          "total_spending": {
            "title": "Total Spending",
            "type": "number"
          },
          "year_month": {
            "title": "Year Month",
            "type": "string"
          }
        },
        "required": [
          "year_month",
          "total_spending",
          "spending_count",
          "deposit_total",
          "deposit_count",
          "by_category",
          "by_company",
          "deposits_by_company",
          "top_categories"
        ],
        "title": "MonthSummary",
        "type": "object"
      },
...
      "SummaryComparisonResponse": {
        "properties": {
          "current": {
            "$ref": "#/components/schemas/MonthSummary"
          },
          "delta_amount": {
            "title": "Delta Amount",
            "type": "number"
          },
          "delta_percent": {
            "title": "Delta Percent",
            "type": "number"
          },
          "pace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonthPaceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "previous": {
            "$ref": "#/components/schemas/MonthSummary"
          }
        },
        "required": [
          "current",
          "previous",
          "delta_amount",
          "delta_percent"
        ],
        "title": "SummaryComparisonResponse",
        "type": "object"
      },
...
      "TrendMonthEntry": {
        "properties": {
          "by_category": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "title": "By Category",
            "type": "object"
          },
          "spending_count": {
            "title": "Spending Count",
            "type": "integer"
          },
          "total_spending": {
            "title": "Total Spending",
            "type": "number"
          },
          "year_month": {
            "title": "Year Month",
            "type": "string"
          }
        },
        "required": [
          "year_month",
          "total_spending",
          "spending_count",
          "by_category"
        ],
        "title": "TrendMonthEntry",
        "type": "object"
      },
      "TrendResponse": {
        "properties": {
          "months": {
            "items": {
              "$ref": "#/components/schemas/TrendMonthEntry"
            },
            "title": "Months",
            "type": "array"
          }
        },
        "required": [
          "months"
        ],
        "title": "TrendResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal finance dashboard API. Ingests bank transaction emails, categorizes them (with optional OpenAI enrichment), and exposes transactions, budgets, summaries, and AI spending insights over a consistent JSON HTTP interface.\n\n**Storage:** dual-backend \u2014 DynamoDB (AWS) or SQLite (local/demo). Selection is driven by `data/config.json`. All endpoints return identical shapes regardless of backend.\n\n**Versioning:** routes are prefixed with `/api/v1/`. Breaking changes will ship as `/api/v2/` rather than in-place.\n\n**Demo:** this is a read-only snapshot of the Tidings demo journal \u2014 the fictional Mira Lin Chen household, covering 2025-05 through 2026-03. All data is invented. Any request other than GET, HEAD, or OPTIONS returns 405; self-host Tidings for a writable journal with bearer-token auth.",
    "title": "Tidings demo API (read-only)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/journal": {
      "get": {
        "operationId": "getJournal",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Day-grouped transaction timeline for a month",
        "tags": [
          "journal"
        ]
      }
    },
    "/api/v1/journal/summaries": {
      "get": {
        "operationId": "listDaySummaries",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaySummariesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List saved AI day summaries for a month",
        "tags": [
          "daily-summaries"
        ]
      }
    },
...
    "/api/v1/summary": {
      "get": {
        "operationId": "getMonthlySummary",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Monthly spending summary with previous-month comparison",
        "tags": [
          "summary"
        ]
      }
    },
    "/api/v1/summary/trend": {
      "get": {
        "operationId": "getSpendingTrend",
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 12,
              "minimum": 2,
              "title": "Months",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Multi-month spending trend (2-12 months)",
        "tags": [
          "summary"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gettidings.com/demo"
    }
  ]
}